# openapi/spec/FEM.openapi.json
{"openapi":"3.0.1","info":{"title":"FEM","description":"","version":"1.0.0"},"tags":[{"name":"FEM"},{"name":"FEM/ユーザ管理"}],"paths":{"/user/insert":{"post":{"summary":"ユーザ登録API","deprecated":false,"description":"ユーザ管理テーブルにレコードを登録するAPI","tags":["FEM/ユーザ管理"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"string","description":"ユーザID"},"userPass":{"type":"string","description":"パスワード（登録時はハッシュ値に変換）"},"userNm":{"type":"string","description":"ユーザ名（姓+全角スペース+名）"},"userMail":{"type":"string","description":"メールアドレス"}},"required":["userId","userPass","userNm","userMail"]}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"msgId":{"type":"string","description":"メッセージID"},"msg":{"type":"string","description":"メッセージ"},"result":{"type":"array","items":{"type":"object","properties":{}},"description":"処理結果"}},"required":["msgId","msg","result"]},"examples":{"1":{"summary":"成功例","value":{"msgId":"info_001_001","msg":"登録が成功しました。","result":[]}}}}},"headers":{}},"500":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"msg":{"type":"string","description":"メッセージ"},"result":{"type":"array","items":{"type":"object","properties":{}},"description":"処理結果"},"msgId":{"type":"string","description":"メッセージID"}},"required":["msgId","msg","result"]},"examples":{"1":{"summary":"異常例","value":{"msgId":"error_001_001","msg":"エラーが発生しました。","result":[]}}}}},"headers":{}}},"security":[]}},"/user/auth":{"get":{"summary":"ユーザ認証API","deprecated":false,"description":"ユーザ管理テーブルから該当するユーザ認証確認を行うAPI","tags":["FEM/ユーザ管理"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"string","description":"ユーザID"},"userPass":{"type":"string","description":"パスワード（登録時はハッシュ値に変換）"}},"required":["userId","userPass"]}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"msgId":{"type":"string","description":"メッセージID"},"msg":{"type":"string","description":"メッセージ"},"result":{"type":"array","items":{"type":"object","properties":{}},"description":"処理結果"}},"required":["msgId","msg","result"]},"examples":{"1":{"summary":"成功例","value":{"msgId":"info_001_002","msg":"認証が成功しました。","result":[]}}}}},"headers":{}},"500":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"msg":{"type":"string","description":"メッセージ"},"result":{"type":"array","items":{"type":"object","properties":{}},"description":"処理結果"},"msgId":{"type":"string","description":"メッセージID"}},"required":["msgId","msg","result"]},"examples":{"1":{"summary":"異常例","value":{"msgId":"error_001_001","msg":"エラーが発生しました。","result":[]}}}}},"headers":{}}},"security":[]}},"/version":{"get":{"summary":"appバージョン","deprecated":false,"description":"アプリケーションのバージョンを返却","tags":["FEM"],"parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"version":{"type":"string"}},"required":["version"]},"examples":{"1":{"summary":"成功例","value":"{\n    version: \"0.0.1\"\n}"}}}},"headers":{}}},"security":[]}}},"components":{"schemas":{"users":{"type":"object","properties":{"sysNo":{"type":"string","description":"システム番号（払い出し）"},"insertDt":{"type":"string","description":"登録日時（yyyymmddHHmmss形式）"},"updateDt":{"type":"string","description":"更新日時（yyyymmddHHmmss形式）"},"insertUser":{"type":"string","description":"登録ユーザ"},"updateUser":{"type":"string","description":"更新ユーザ"},"userId":{"type":"string","description":"ユーザID"},"userPass":{"type":"string","description":"パスワード（登録時はハッシュ値に変換）"},"userNm":{"type":"string","description":"ユーザ名（姓+全角スペース+名）"},"userMail":{"type":"string","description":"メールアドレス"}},"required":["userPass","sysNo","userMail","userId","userNm","insertUser","updateUser","insertDt","updateDt"]},"friends":{"type":"object","properties":{"sysNo":{"type":"string","description":"システム番号（払い出し）"},"insertDt":{"type":"string","description":"登録日時（yyyymmddHHmmss形式）"},"updateDt":{"type":"string","description":"更新日時（yyyymmddHHmmss形式）"},"insertUser":{"type":"string","description":"登録ユーザ"},"updateUser":{"type":"string","description":"更新ユーザ"},"friendNm":{"type":"string","description":"友達名"}},"required":["sysNo","insertDt","updateDt","insertUser","updateUser","friendNm"]},"events":{"type":"object","properties":{"sysNo":{"type":"string","description":"システム番号（払い出し）"},"insertDt":{"type":"string","description":"登録日時（yyyymmddHHmmss形式）"},"updateDt":{"type":"string","description":"更新日時（yyyymmddHHmmss形式）"},"insertUser":{"type":"string","description":"登録ユーザ"},"updateUser":{"type":"string","description":"更新ユーザ"},"eventId":{"type":"string","description":"イベントID番号（払い出し）"},"eventNm":{"type":"string","description":"イベント名"},"eventStartDate":{"type":"string","description":"イベント開始日（yyyymmdd形式）"},"eventEndDate":{"type":"string","description":"イベント終了日（yyyymmdd形式）"},"eventStartTime":{"type":"string","description":"イベント開始時間（HHmmss形式）"},"eventEndTime":{"type":"string","description":"イベント終了時間（HHmmss形式）"}},"required":["sysNo","eventNm","eventId","eventStartDate","eventEndDate","insertDt","updateDt","insertUser","updateUser"]}},"securitySchemes":{}},"servers":[]}